services/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-text-origin-risk-api"
version = "0.1.0"
description = "FastAPI service for cautious AI text origin risk analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12,<5.0",
  "fastapi>=0.115,<0.116",
  "langdetect>=1.0.9,<2.0",
  "pydantic-settings>=2.6,<3.0",
  "pypdf>=5.1,<6.0",
  "python-docx>=1.1,<2.0",
  "python-multipart>=0.0.17,<0.1",
  "striprtf>=0.0.26,<0.1",
  "uvicorn[standard]>=0.32,<0.33"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28,<0.29",
  "pytest>=8.3,<9.0"
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
